move commands outside the lb folder, add sde_download to update_sde
